Add --color always
Usage in a GitHub Action doesn't appear as a terminal device, but color
escapes are fully supported. This fools the CLI into colorizing anyway.

The option is recent (as of v1.3.0.1), and users are free to choose any
platform-version, so we'll make this conditional on use of that option
succeeding in an innocuous subcommand so as not to break anyone pinned
on an older version.