Fail on empty Slack webhook URL
Currently, if the `slack-webhook` input is a secret that does not exist in the repository, the notify steps will be skipped. Since this is usually a sign of misconfiguration, the action should produce a failure instead. However, since the inputs in this scenario look identical to how a user would opt to not have the action notify, the input is changed to optional and the notify steps are instead skipped when matching on a default input value, rather than an empty value.

For example, there is no `SLACK_WEBHOOK_URL` in the repository secrets, `secrets.SLACK_WEBHOOK_URL` is an empty value that replaces the default value of the `slack-webhook` input. `rtCamp/action-slack-notify` will then [throw an error][0] attempting to use this as a webhook.

[0]: https://github.com/rtCamp/action-slack-notify/blob/master/main.go#L60